게시글 작성시 loginUser로 user 값 입력

+++ src/main/java/kernel360/devinside/domain/post/service/PostService.java
@@ -2,12 +2,15 @@ package kernel360.devinside.domain.post.service;
 
 
 import jakarta.transaction.Transactional;
-import kernel360.devinside.domain.post.domain.*;
+import kernel360.devinside.domain.post.domain.Post;
+import kernel360.devinside.domain.post.domain.PostCategory;
 import kernel360.devinside.domain.post.dto.Pagination;
 import kernel360.devinside.domain.post.dto.PostRequest;
 import kernel360.devinside.domain.post.dto.PostResponse;
 import kernel360.devinside.domain.post.exception.PostNotFoundException;
 import kernel360.devinside.domain.post.repository.PostRepository;
+import kernel360.devinside.domain.user.domain.LoginUser;
+import kernel360.devinside.domain.user.domain.User;
 import kernel360.devinside.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -23,16 +26,19 @@ public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
-    public Post create(PostRequest postRequest) {
+    public Post create(
+            PostRequest postRequest,
+            LoginUser loginUser
+    ) {
 
-//         Optional<User> user = userRepository.findById(id);
-//         if(user.isPresent()) { }
+        User user = userRepository.findById(loginUser.id())
+                .orElseThrow(() -> new RuntimeException("존재하지 않는 사용자 입니다."));
 
         Post post = Post.builder()
                 .postCategory(postRequest.getPostCategory())
                 .title(postRequest.getTitle())
                 .content(postRequest.getContent())
-//                .user(user)
+                .user(user)
                 .view(0)
                 .liked(false)
                 .hated(false)
@@ -82,22 +88,43 @@ public class PostService {
         return post;
     }
 
-    public void delete(Long id) {
+
+    public void delete(LoginUser loginUser, Long id) {
+
+        if (loginUser == null) {
+            throw new RuntimeException("로그인 정보가 없습니다.");
+        }
 
         Post post = postRepository.findById(id)
+                .map(p -> {
+                    if (p.getUser().getId().equals(loginUser.id())) {
+                        return p;
+                    } else {
+                        throw new RuntimeException("삭제 권한이 없습니다.");
+                    }
+                })
                 .orElseThrow(() -> new PostNotFoundException("삭제하려는 게시글이 존재하지 않습니다."));
 
         postRepository.delete(post);
 
     }
 
-    public Post update(Long id, PostRequest postRequest) {
+    public Post update(LoginUser loginUser, Long id, PostRequest postRequest) {
+
+        if (loginUser == null) {
+            throw new RuntimeException("로그인 정보가 없습니다.");
+        }
+
 
         return postRepository.findById(id).map(post -> {
-            post.updateContent(postRequest.getContent());
-            post.updateTitle(postRequest.getTitle());
-            postRepository.save(post);
-            return post;
+            if (post.getUser().getId().equals(loginUser.id())) {
+                post.updateContent(postRequest.getContent());
+                post.updateTitle(postRequest.getTitle());
+                postRepository.save(post);
+                return post;
+            } else {
+                throw new RuntimeException("수정 권한이 없습니다.");
+            }
         }).orElseThrow(() -> {
             return new PostNotFoundException("수정하려는 게시글이 존재하지 않습니다.");
         });
